@@ -16,66 +16,69 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.12",
-    "@angular/cdk": "~13.3.9",
-    "@angular/common": "~13.3.12",
-    "@angular/compiler": "~13.3.12",
-    "@angular/core": "~13.3.12",
-    "@angular/forms": "~13.3.12",
-    "@angular/localize": "~13.3.12",
-    "@angular/platform-browser": "~13.3.12",
-    "@angular/platform-browser-dynamic": "~13.3.12",
-    "@angular/pwa": "~13.3.10",
-    "@angular/router": "~13.3.12",
-    "@angular/service-worker": "~13.3.12",
-    "@handsontable/angular": "^12.3.0",
-    "@ng-bootstrap/ng-bootstrap": "^11.0.1",
-    "@tinymce/tinymce-angular": "^6.0.1",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/fire": "^7.5.0",
+    "@angular/forms": "^14.2.12",
+    "@angular/localize": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/pwa": "^14.2.10",
+    "@angular/router": "^14.2.12",
+    "@angular/service-worker": "^14.2.12",
+    "@handsontable/angular": "^12.3.1",
+    "@ng-bootstrap/ng-bootstrap": "^13.1.1",
+    "@popperjs/core": "~2.10.2",
+    "@tinymce/tinymce-angular": "^7.0.0",
     "d3": "^7.8.2",
     "file-saver": "^2.0.5",
-    "handsontable": "^12.3.0",
+    "firebase": "^9.17.2",
+    "handsontable": "^12.3.1",
     "intersection-observer": "^0.12.2",
     "moment-timezone": "^0.5.40",
-    "ng-dynamic-component": "~10.1.0",
+    "ng-dynamic-component": "^10.6.1",
     "ng-in-viewport": "^6.1.5",
-    "ngx-captcha": "^11.0.0",
-    "ngx-page-scroll-core": "^8.0.0",
+    "ngx-captcha": "^12.0.2",
+    "ngx-page-scroll-core": "^9.0.0",
     "rxjs": "~7.5.7",
-    "tslib": "^2.4.1",
-    "ua-parser-js": "^0.7.32",
+    "tslib": "^2.5.0",
+    "ua-parser-js": "^0.7.33",
     "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^13.0.4",
-    "@angular-devkit/build-angular": "~13.3.10",
-    "@angular-eslint/builder": "~13.2.1",
-    "@angular-eslint/eslint-plugin": "~13.2.1",
-    "@angular-eslint/eslint-plugin-template": "~13.2.1",
-    "@angular-eslint/schematics": "~13.2.1",
-    "@angular-eslint/template-parser": "~13.2.1",
-    "@angular/cli": "^13.3.10",
-    "@angular/compiler-cli": "~13.3.12",
+    "@angular-builders/jest": "^14.1.0",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular-eslint/builder": "^14.4.0",
+    "@angular-eslint/eslint-plugin": "^14.4.0",
+    "@angular-eslint/eslint-plugin-template": "^14.4.0",
+    "@angular-eslint/schematics": "^14.4.0",
+    "@angular-eslint/template-parser": "^14.4.0",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
     "@types/d3": "^7.4.0",
-    "@types/jest": "^27.5.2",
-    "@typescript-eslint/eslint-plugin": "^5.47.0",
-    "@typescript-eslint/parser": "^5.47.0",
-    "eslint": "^8.30.0",
+    "@types/jest": "^28.1.8",
+    "@typescript-eslint/eslint-plugin": "^5.52.0",
+    "@typescript-eslint/parser": "^5.52.0",
+    "eslint": "^8.34.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-plugin-deprecation": "^1.3.3",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.1.7",
-    "eslint-plugin-jsdoc": "^39.6.4",
+    "eslint-plugin-jest": "^27.2.1",
+    "eslint-plugin-jsdoc": "^40.0.0",
     "eslint-plugin-rxjs": "^5.0.2",
-    "jest": "^27.5.1",
+    "jest": "^28.1.3",
     "lintspaces-cli": "^0.7.1",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.20",
-    "rimraf": "^3.0.2",
+    "postcss": "^8.4.21",
+    "rimraf": "^4.1.2",
     "run-script-os": "^1.1.6",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
-    "typescript": "~4.6.4",
-    "webpack-bundle-analyzer": "^4.7.0"
+    "typescript": "~4.7.4",
+    "webpack-bundle-analyzer": "^4.8.0"
   }
 }
