@@ -43,9 +43,16 @@ import teammates.test.ThreadHelper;
  */
 public abstract class BaseE2ETestCase extends BaseTestCaseWithDatabaseAccess {
 
-    static final BackDoor BACKDOOR = BackDoor.getInstance();
+    /**
+     * Backdoor used to call APIs.
+     */
+    protected static final BackDoor BACKDOOR = BackDoor.getInstance();
 
-    DataBundle testData;
+    /**
+     * Data to be used in the test.
+     */
+    protected DataBundle testData;
+
     private Browser browser;
 
     @BeforeClass
@@ -295,7 +302,10 @@ public abstract class BaseE2ETestCase extends BaseTestCaseWithDatabaseAccess {
         return getInstructor(instructor.getCourseId(), instructor.getEmail());
     }
 
-    String getKeyForInstructor(String courseId, String instructorEmail) {
+    /**
+     * Gets registration key for a given instructor.
+     */
+    protected String getKeyForInstructor(String courseId, String instructorEmail) {
         return getInstructor(courseId, instructorEmail).getKey();
     }
 
@@ -304,7 +314,10 @@ public abstract class BaseE2ETestCase extends BaseTestCaseWithDatabaseAccess {
         return BACKDOOR.getStudent(student.getCourse(), student.getEmail());
     }
 
-    String getKeyForStudent(StudentAttributes student) {
+    /**
+     * Gets registration key for a given student.
+     */
+    protected String getKeyForStudent(StudentAttributes student) {
         return getStudent(student).getKey();
     }
 
