@@ -103,24 +103,28 @@ public class FeedbackResultsPage extends AppPage {
         }
     }
 
+    public void verifyQuestionHasResponsesNotVisibleForPreview(int questionNum) {
+        verifyQuestionDoesNotShowResponses(questionNum);
+        verifyNonVisibleResponseAlertPresent(questionNum);
+    }
+
+    public void verifyQuestionHasCommentsNotVisibleForPreview(int questionNum, List<String> commentsNotVisible) {
+        verifyQuestionDoesNotShowComments(questionNum, commentsNotVisible);
+        verifyNonVisibleCommentAlertPresent(questionNum);
+    }
+
     public void verifyNumScaleStatistics(int questionNum, String[] expectedStats) {
         verifyTableRowValues(getNumScaleStatistics(questionNum), expectedStats);
     }
 
-    public void verifyRubricStatistics(int questionNum, String[][] expectedStats, String[][] expectedStatsExcludingSelf,
-                                       String[][] expectedStatsPerRecipient) {
+    public void verifyRubricStatistics(int questionNum, String[][] expectedStats,
+                                       String[][] expectedStatsExcludingSelf) {
         WebElement excludeSelfCheckbox = getRubricExcludeSelfCheckbox(questionNum);
         markOptionAsUnselected(excludeSelfCheckbox);
         verifyTableBodyValues(getRubricStatistics(questionNum), expectedStats);
 
         markOptionAsSelected(excludeSelfCheckbox);
         verifyTableBodyValues(getRubricStatistics(questionNum), expectedStatsExcludingSelf);
-
-        sortRubricPerRecipientStatsPerCriterion(questionNum, 2);
-        verifyTableBodyValues(getRubricPerRecipientStatsPerCriterion(questionNum), expectedStatsPerRecipient);
-
-        sortRubricPerRecipientStatsOverall(questionNum, 2);
-        verifyTableBodyValues(getRubricPerRecipientStatsPerCriterion(questionNum), expectedStatsPerRecipient);
     }
 
     public void verifyContributionStatistics(int questionNum, String[] expectedStats) {
@@ -140,6 +144,47 @@ public class FeedbackResultsPage extends AppPage {
         }
         if (!commentEditor.isEmpty()) {
             assertEquals(commentEditor, getCommentEditor(commentField));
+        }
+    }
+
+    private void verifyQuestionDoesNotShowResponses(int questionNum) {
+        WebElement questionResponsesSection = getQuestionResponsesSection(questionNum);
+        try {
+            questionResponsesSection.findElement(By.className("all-responses"));
+            fail("Question " + questionNum + " should not display any actual responses when previewing results.");
+        } catch (NoSuchElementException e) {
+            // success
+        }
+    }
+
+    private void verifyQuestionDoesNotShowComments(int questionNum, List<String> commentsNotVisible) {
+        List<WebElement> commentsOfQuestion = getCommentFields(questionNum);
+        for (String commentString : commentsNotVisible) {
+            for (WebElement comment : commentsOfQuestion) {
+                if (comment.findElement(By.className("comment-text")).getText().equals(commentString)) {
+                    fail("Comment \"" + commentString + "\" should not be present in question " + questionNum);
+                }
+            }
+        }
+    }
+
+    private void verifyNonVisibleResponseAlertPresent(int questionNum) {
+        WebElement questionResponsesSection = getQuestionResponsesSection(questionNum);
+        try {
+            questionResponsesSection.findElement(By.className("non-visible-response-alert"));
+        } catch (NoSuchElementException e) {
+            fail("Question " + questionNum
+                    + " should display an alert message for hidden responses when previewing results.");
+        }
+    }
+
+    private void verifyNonVisibleCommentAlertPresent(int questionNum) {
+        WebElement questionResponsesSection = getQuestionResponsesSection(questionNum);
+        try {
+            questionResponsesSection.findElement(By.className("non-visible-comment-alert"));
+        } catch (NoSuchElementException e) {
+            fail("Question " + questionNum
+                    + " should display an alert message for hidden comments when previewing results.");
         }
     }
 
@@ -313,7 +358,7 @@ public class FeedbackResultsPage extends AppPage {
     }
 
     private String getQuestionText(int questionNum) {
-        return getQuestionResponsesSection(questionNum).findElement(By.id("question-text")).getText().trim();
+        return getQuestionResponsesSection(questionNum).findElement(By.className("question-text")).getText().trim();
     }
 
     private String getMcqAddInfo(FeedbackMcqQuestionDetails questionDetails) {
@@ -406,7 +451,8 @@ public class FeedbackResultsPage extends AppPage {
     }
 
     private void showAdditionalInfo(int qnNumber) {
-        WebElement additionalInfoLink = getQuestionResponsesSection(qnNumber).findElement(By.id("additional-info-link"));
+        WebElement additionalInfoLink =
+                getQuestionResponsesSection(qnNumber).findElement(By.className("additional-info-button"));
         if ("[more]".equals(additionalInfoLink.getText())) {
             click(additionalInfoLink);
             waitUntilAnimationFinish();
@@ -415,7 +461,7 @@ public class FeedbackResultsPage extends AppPage {
 
     private String getAdditionalInfo(int questionNum) {
         showAdditionalInfo(questionNum);
-        return getQuestionResponsesSection(questionNum).findElement(By.id("additional-info")).getText();
+        return getQuestionResponsesSection(questionNum).findElement(By.className("additional-info")).getText();
     }
 
     private WebElement getGivenResponseField(int questionNum, String receiver) {
@@ -532,24 +578,8 @@ public class FeedbackResultsPage extends AppPage {
         return getQuestionResponsesSection(questionNum).findElement(By.id("rubric-statistics"));
     }
 
-    private WebElement getRubricPerRecipientStatsPerCriterion(int questionNum) {
-        return getQuestionResponsesSection(questionNum).findElement(By.id("rubric-recipient-statistics-per-criterion"));
-    }
-
-    private void sortRubricPerRecipientStatsPerCriterion(int questionNum, int colNum) {
-        click(getRubricPerRecipientStatsPerCriterion(questionNum).findElements(By.tagName("th")).get(colNum - 1));
-    }
-
-    private WebElement getRubricPerRecipientStatsOverall(int questionNum) {
-        return getQuestionResponsesSection(questionNum).findElement(By.id("rubric-recipient-statistics-overall"));
-    }
-
-    private void sortRubricPerRecipientStatsOverall(int questionNum, int colNum) {
-        click(getRubricPerRecipientStatsOverall(questionNum).findElements(By.tagName("th")).get(colNum - 1));
-    }
-
     private boolean isCommentByResponseGiver(WebElement commentField) {
-        return commentField.findElements(By.id("by-response-giver")).size() > 0;
+        return commentField.findElements(By.className("by-response-giver")).size() > 0;
     }
 
     private String getCommentGiver(WebElement commentField) {
@@ -569,7 +599,7 @@ public class FeedbackResultsPage extends AppPage {
     private WebElement getCommentField(int questionNum, String commentString) {
         List<WebElement> commentFields = getCommentFields(questionNum);
         for (WebElement comment : commentFields) {
-            if (comment.findElement(By.id("comment-text")).getText().equals(commentString)) {
+            if (comment.findElement(By.className("comment-text")).getText().equals(commentString)) {
                 return comment;
             }
         }
