@@ -111,7 +111,7 @@ public class InstructorSearchPage extends AppPage {
 
     private WebElement getDeleteButton(CourseAttributes course, String studentEmail) {
         WebElement studentRow = getStudentRow(course, studentEmail);
-        return studentRow.findElement(By.id("btn-delete"));
+        return studentRow.findElement(By.cssSelector("[id^='btn-delete-']"));
     }
 
     private WebElement getStudentRow(CourseAttributes course, String studentEmail) {
@@ -132,7 +132,7 @@ public class InstructorSearchPage extends AppPage {
 
     public InstructorCourseStudentDetailsViewPage clickViewStudent(CourseAttributes course, String studentEmail) {
         WebElement studentRow = getStudentRow(course, studentEmail);
-        WebElement viewButton = studentRow.findElement(By.id("btn-view-details"));
+        WebElement viewButton = studentRow.findElement(By.cssSelector("[id^='btn-view-details-']"));
         click(viewButton);
         ThreadHelper.waitFor(2000);
         switchToNewWindow();
@@ -141,7 +141,7 @@ public class InstructorSearchPage extends AppPage {
 
     public InstructorCourseStudentDetailsEditPage clickEditStudent(CourseAttributes course, String studentEmail) {
         WebElement studentRow = getStudentRow(course, studentEmail);
-        WebElement viewButton = studentRow.findElement(By.id("btn-edit-details"));
+        WebElement viewButton = studentRow.findElement(By.cssSelector("[id^='btn-edit-details-']"));
         click(viewButton);
         ThreadHelper.waitFor(2000);
         switchToNewWindow();
@@ -150,7 +150,7 @@ public class InstructorSearchPage extends AppPage {
 
     public InstructorStudentRecordsPage clickViewAllRecords(CourseAttributes course, String studentEmail) {
         WebElement studentRow = getStudentRow(course, studentEmail);
-        WebElement viewButton = studentRow.findElement(By.id("btn-view-records"));
+        WebElement viewButton = studentRow.findElement(By.cssSelector("[id^='btn-view-records-']"));
         click(viewButton);
         ThreadHelper.waitFor(2000);
         switchToNewWindow();
