@@ -27,7 +27,8 @@ public enum EmailType {
     INSTRUCTOR_COURSE_REJOIN_AFTER_GOOGLE_ID_RESET("TEAMMATES: Your account has been reset for course [%s][Course ID: %s]"),
     USER_COURSE_REGISTER("TEAMMATES: Registered for Course [%s][Course ID: %s]"),
     SEVERE_LOGS_COMPILATION("TEAMMATES (%s): Severe Error Logs Compilation"),
-    SESSION_LINKS_RECOVERY("TEAMMATES: Access links for recent feedback sessions");
+    SESSION_LINKS_RECOVERY("TEAMMATES: Access links for recent feedback sessions"),
+    LOGIN("TEAMMATES: Log in to TEAMMATES");
 
     private final String subject;
 
