@@ -74,21 +74,29 @@ abstract class BasicFeedbackSubmissionAction extends Action {
                     student.getSection(),
                     Const.InstructorPermissions.CAN_MODIFY_SESSION_COMMENT_IN_SECTIONS);
         } else if (!StringHelper.isEmpty(previewAsPerson)) {
-            gateKeeper.verifyLoggedInUserPrivileges(userInfo);
-            gateKeeper.verifyAccessible(
-                    logic.getInstructorForGoogleId(feedbackSession.getCourseId(), userInfo.getId()), feedbackSession,
-                    Const.InstructorPermissions.CAN_MODIFY_SESSION);
+            checkAccessControlForPreview(feedbackSession, false);
         } else {
             gateKeeper.verifyAccessible(student, feedbackSession);
-            if (!StringHelper.isEmpty(student.getGoogleId())) {
-                if (userInfo == null) {
-                    // Student is associated to a google ID; even if registration key is passed, do not allow access
-                    throw new UnauthorizedAccessException("Login is required to access this feedback session");
-                } else if (!userInfo.id.equals(student.getGoogleId())) {
-                    // Logged in student is not the same as the student registered for the given key, do not allow access
-                    throw new UnauthorizedAccessException("You are not authorized to access this feedback session");
-                }
-            }
+            verifyMatchingGoogleId(student.getGoogleId());
+        }
+    }
+
+    /**
+     * Checks the access control for student feedback result.
+     */
+    void checkAccessControlForStudentFeedbackResult(
+            StudentAttributes student, FeedbackSessionAttributes feedbackSession) throws UnauthorizedAccessException {
+        if (student == null) {
+            throw new UnauthorizedAccessException("Trying to access system using a non-existent student entity");
+        }
+
+        String previewAsPerson = getRequestParamValue(Const.ParamsNames.PREVIEWAS);
+
+        if (StringHelper.isEmpty(previewAsPerson)) {
+            gateKeeper.verifyAccessible(student, feedbackSession);
+            verifyMatchingGoogleId(student.getGoogleId());
+        } else {
+            checkAccessControlForPreview(feedbackSession, false);
         }
     }
 
@@ -125,21 +133,57 @@ abstract class BasicFeedbackSubmissionAction extends Action {
             gateKeeper.verifyAccessible(logic.getInstructorForGoogleId(feedbackSession.getCourseId(), userInfo.getId()),
                     feedbackSession, Const.InstructorPermissions.CAN_MODIFY_SESSION_COMMENT_IN_SECTIONS);
         } else if (!StringHelper.isEmpty(previewAsPerson)) {
-            gateKeeper.verifyLoggedInUserPrivileges(userInfo);
-            gateKeeper.verifyAccessible(logic.getInstructorForGoogleId(feedbackSession.getCourseId(), userInfo.getId()),
-                    feedbackSession, Const.InstructorPermissions.CAN_MODIFY_SESSION);
+            checkAccessControlForPreview(feedbackSession, true);
         } else {
             gateKeeper.verifySessionSubmissionPrivilegeForInstructor(feedbackSession, instructor);
-            if (!StringHelper.isEmpty(instructor.getGoogleId())) {
-                if (userInfo == null) {
-                    // Instructor is associated to a google ID; even if registration key is passed, do not allow access
-                    throw new UnauthorizedAccessException("Login is required to access this feedback session");
-                } else if (!userInfo.id.equals(instructor.getGoogleId())) {
-                    // Logged in instructor is not the same as the instructor registered for the given key,
-                    // do not allow access
-                    throw new UnauthorizedAccessException("You are not authorized to access this feedback session");
-                }
+            verifyMatchingGoogleId(instructor.getGoogleId());
+        }
+    }
+
+    /**
+     * Checks the access control for instructor feedback result.
+     */
+    void checkAccessControlForInstructorFeedbackResult(
+            InstructorAttributes instructor, FeedbackSessionAttributes feedbackSession) throws UnauthorizedAccessException {
+        if (instructor == null) {
+            throw new UnauthorizedAccessException("Trying to access system using a non-existent instructor entity");
+        }
+
+        String previewAsPerson = getRequestParamValue(Const.ParamsNames.PREVIEWAS);
+
+        if (StringHelper.isEmpty(previewAsPerson)) {
+            gateKeeper.verifyAccessible(instructor, feedbackSession,
+                    Const.InstructorPermissions.CAN_VIEW_SESSION_IN_SECTIONS);
+            verifyMatchingGoogleId(instructor.getGoogleId());
+        } else {
+            checkAccessControlForPreview(feedbackSession, true);
+        }
+    }
+
+    private void verifyMatchingGoogleId(String googleId) throws UnauthorizedAccessException {
+        if (!StringHelper.isEmpty(googleId)) {
+            if (userInfo == null) {
+                // Student/Instructor is associated to a google ID; even if registration key is passed, do not allow access
+                throw new UnauthorizedAccessException("Login is required to access this feedback session");
+            } else if (!userInfo.id.equals(googleId)) {
+                // Logged in student/instructor is not the same as the student/instructor registered for the given key,
+                // do not allow access
+                throw new UnauthorizedAccessException("You are not authorized to access this feedback session");
             }
+        }
+    }
+
+    private void checkAccessControlForPreview(FeedbackSessionAttributes feedbackSession, boolean isInstructor)
+            throws UnauthorizedAccessException {
+        gateKeeper.verifyLoggedInUserPrivileges(userInfo);
+        if (isInstructor) {
+            gateKeeper.verifyAccessible(
+                    logic.getInstructorForGoogleId(feedbackSession.getCourseId(), userInfo.getId()), feedbackSession,
+                    Const.InstructorPermissions.CAN_MODIFY_SESSION);
+        } else {
+            gateKeeper.verifyAccessible(
+                    logic.getInstructorForGoogleId(feedbackSession.getCourseId(), userInfo.getId()), feedbackSession,
+                    Const.InstructorPermissions.CAN_MODIFY_SESSION);
         }
     }
 
