@@ -6,13 +6,14 @@ import teammates.common.datatransfer.attributes.FeedbackSessionAttributes;
 import teammates.common.datatransfer.attributes.InstructorAttributes;
 import teammates.common.datatransfer.attributes.StudentAttributes;
 import teammates.common.util.Const;
+import teammates.common.util.StringHelper;
 import teammates.ui.output.SessionResultsData;
 import teammates.ui.request.Intent;
 
 /**
  * Gets feedback session results including statistics where necessary.
  */
-class GetSessionResultsAction extends Action {
+class GetSessionResultsAction extends BasicFeedbackSubmissionAction {
 
     @Override
     AuthType getMinAuthLevel() {
@@ -26,6 +27,8 @@ class GetSessionResultsAction extends Action {
 
         FeedbackSessionAttributes fs = getNonNullFeedbackSession(feedbackSessionName, courseId);
         Intent intent = Intent.valueOf(getNonNullRequestParamValue(Const.ParamsNames.INTENT));
+        String previewAsPerson = getRequestParamValue(Const.ParamsNames.PREVIEWAS);
+        boolean isPreviewResults = !StringHelper.isEmpty(previewAsPerson);
         switch (intent) {
         case FULL_DETAIL:
             gateKeeper.verifyLoggedInUserPrivileges(userInfo);
@@ -33,18 +36,18 @@ class GetSessionResultsAction extends Action {
             gateKeeper.verifyAccessible(instructor, fs);
             break;
         case INSTRUCTOR_RESULT:
-            instructor = getPossiblyUnregisteredInstructor(courseId);
-            gateKeeper.verifyAccessible(instructor, fs);
-            if (!fs.isPublished()) {
+            if (!isPreviewResults && !fs.isPublished()) {
                 throw new UnauthorizedAccessException("This feedback session is not yet published.", true);
             }
+            instructor = getInstructorOfCourseFromRequest(courseId);
+            checkAccessControlForInstructorFeedbackResult(instructor, fs);
             break;
         case STUDENT_RESULT:
-            StudentAttributes student = getPossiblyUnregisteredStudent(courseId);
-            gateKeeper.verifyAccessible(student, fs);
-            if (!fs.isPublished()) {
+            if (!isPreviewResults && !fs.isPublished()) {
                 throw new UnauthorizedAccessException("This feedback session is not yet published.", true);
             }
+            StudentAttributes student = getStudentOfCourseFromRequest(courseId);
+            checkAccessControlForStudentFeedbackResult(student, fs);
             break;
         case INSTRUCTOR_SUBMISSION:
         case STUDENT_SUBMISSION:
@@ -65,6 +68,9 @@ class GetSessionResultsAction extends Action {
         FeedbackResultFetchType fetchType = FeedbackResultFetchType.parseFetchType(
                 getRequestParamValue(Const.ParamsNames.FEEDBACK_RESULTS_SECTION_BY_GIVER_RECEIVER));
 
+        String previewAsPerson = getRequestParamValue(Const.ParamsNames.PREVIEWAS);
+        boolean isPreviewResults = !StringHelper.isEmpty(previewAsPerson);
+
         SessionResultsBundle bundle;
         InstructorAttributes instructor;
         StudentAttributes student;
@@ -78,10 +84,10 @@ class GetSessionResultsAction extends Action {
             return new JsonResult(SessionResultsData.initForInstructor(bundle));
         case INSTRUCTOR_RESULT:
             // Section name filter is not applicable here
-            instructor = getPossiblyUnregisteredInstructor(courseId);
+            instructor = getInstructorOfCourseFromRequest(courseId);
 
             bundle = logic.getSessionResultsForUser(feedbackSessionName, courseId, instructor.getEmail(),
-                    true, questionId);
+                    true, questionId, isPreviewResults);
 
             // Build a fake student object, as the results will be displayed as if they are displayed to a student
             student = StudentAttributes.builder(instructor.getCourseId(), instructor.getEmail())
@@ -91,10 +97,10 @@ class GetSessionResultsAction extends Action {
             return new JsonResult(SessionResultsData.initForStudent(bundle, student));
         case STUDENT_RESULT:
             // Section name filter is not applicable here
-            student = getPossiblyUnregisteredStudent(courseId);
+            student = getStudentOfCourseFromRequest(courseId);
 
             bundle = logic.getSessionResultsForUser(feedbackSessionName, courseId, student.getEmail(),
-                    false, questionId);
+                    false, questionId, isPreviewResults);
 
             return new JsonResult(SessionResultsData.initForStudent(bundle, student));
         case INSTRUCTOR_SUBMISSION:
