@@ -1266,7 +1266,7 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
         // Alice will see 3 responses
         SessionResultsBundle bundle = frLogic.getSessionResultsForUser(
                 "First Session", "FQLogicPCT.CS2104", "FQLogicPCT.alice.b@gmail.tmt",
-                false, question.getId());
+                false, question.getId(), false);
         assertEquals(1, bundle.getQuestionResponseMap().size());
         List<FeedbackResponseAttributes> responseForQuestion =
                 bundle.getQuestionResponseMap().entrySet().iterator().next().getValue();
@@ -1275,7 +1275,7 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
         // Benny will see 3 responses
         bundle = frLogic.getSessionResultsForUser(
                 "First Session", "FQLogicPCT.CS2104", "FQLogicPCT.benny.c@gmail.tmt",
-                false, question.getId());
+                false, question.getId(), false);
         assertEquals(1, bundle.getQuestionResponseMap().size());
         responseForQuestion = bundle.getQuestionResponseMap().entrySet().iterator().next().getValue();
         assertEquals(3, responseForQuestion.size());
@@ -1283,7 +1283,7 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
         // Charlie will see 2 responses
         bundle = frLogic.getSessionResultsForUser(
                 "First Session", "FQLogicPCT.CS2104", "FQLogicPCT.charlie.d@gmail.tmt",
-                false, question.getId());
+                false, question.getId(), false);
         assertEquals(1, bundle.getQuestionResponseMap().size());
         responseForQuestion = bundle.getQuestionResponseMap().entrySet().iterator().next().getValue();
         assertEquals(2, responseForQuestion.size());
@@ -1291,7 +1291,7 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
         // Danny will see 2 responses
         bundle = frLogic.getSessionResultsForUser(
                 "First Session", "FQLogicPCT.CS2104", "FQLogicPCT.danny.e@gmail.tmt",
-                false, question.getId());
+                false, question.getId(), false);
         assertEquals(1, bundle.getQuestionResponseMap().size());
         responseForQuestion = bundle.getQuestionResponseMap().entrySet().iterator().next().getValue();
         assertEquals(2, responseForQuestion.size());
@@ -1299,7 +1299,7 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
         // Emily will see 1 response
         bundle = frLogic.getSessionResultsForUser(
                 "First Session", "FQLogicPCT.CS2104", "FQLogicPCT.emily.f@gmail.tmt",
-                false, question.getId());
+                false, question.getId(), false);
         assertEquals(1, bundle.getQuestionResponseMap().size());
         responseForQuestion = bundle.getQuestionResponseMap().entrySet().iterator().next().getValue();
         assertEquals(1, responseForQuestion.size());
@@ -1314,9 +1314,12 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
 
         // Test result bundle for student1
         StudentAttributes student = responseBundle.students.get("student1InCourse1");
+
+        ______TS("student himself/herself is viewing results");
+
         SessionResultsBundle bundle = frLogic.getSessionResultsForUser(
                 session.getFeedbackSessionName(), session.getCourseId(), student.getEmail(),
-                false, null);
+                false, null, false);
 
         // Student can see responses: q1r1, q2r1,3, q3r1, qr4r2-3, q5r1, q7r1-2, q8r1-2
         // We don't check the actual IDs as this is also implicitly tested
@@ -1371,6 +1374,54 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
         // no entry in comment visibility table
         Map<Long, Boolean> commentGiverVisibilityTable = bundle.getCommentGiverVisibilityTable();
         assertEquals(0, commentGiverVisibilityTable.size());
+
+        // preview invisibility info should be empty because this is NOT previewing results
+        assertEquals(0, bundle.getQuestionsNotVisibleForPreviewMap().size());
+        assertEquals(0, bundle.getQuestionsWithCommentNotVisibleForPreview().size());
+
+        ______TS("instructor is previewing results as student");
+
+        bundle = frLogic.getSessionResultsForUser(
+                session.getFeedbackSessionName(), session.getCourseId(), student.getEmail(),
+                false, null, true);
+
+        // instructors can see responses (when previewing results as the student): qr4r2-3
+        totalResponse = 0;
+        for (Map.Entry<String, List<FeedbackResponseAttributes>> entry
+                : bundle.getQuestionResponseMap().entrySet()) {
+            totalResponse += entry.getValue().size();
+        }
+        totalMissingResponse = 0;
+        for (Map.Entry<String, List<FeedbackResponseAttributes>> entry
+                : bundle.getQuestionMissingResponseMap().entrySet()) {
+            totalMissingResponse += entry.getValue().size();
+        }
+        assertEquals(2, totalResponse);
+        // student should not see missing responses
+        assertEquals(0, totalMissingResponse);
+        // only q4 and q6 are visible to instructors previewing the results
+        // but q6 has no viewable response for the student himself/herself
+        assertEquals(1, bundle.getQuestionsMap().size());
+        assertEquals(1, bundle.getQuestionResponseMap().size());
+        assertEquals(1, bundle.getQuestionMissingResponseMap().size());
+
+        // Test the generated response visibilityTable for userNames.
+        responseGiverVisibilityTable = bundle.getResponseGiverVisibilityTable();
+        assertTrue(responseGiverVisibilityTable.get(getResponseId("qn4.resp2", responseBundle)));
+        assertFalse(responseGiverVisibilityTable.get(getResponseId("qn4.resp3", responseBundle)));
+        assertEquals(totalResponse, responseGiverVisibilityTable.size());
+
+        responseRecipientVisibilityTable = bundle.getResponseRecipientVisibilityTable();
+        assertTrue(responseRecipientVisibilityTable.get(getResponseId("qn4.resp2", responseBundle)));
+        assertTrue(responseRecipientVisibilityTable.get(getResponseId("qn4.resp3", responseBundle)));
+        assertEquals(totalResponse, responseRecipientVisibilityTable.size());
+
+        // no entry in comment visibility table
+        assertEquals(0, bundle.getCommentGiverVisibilityTable().size());
+
+        // q1-3, q5, q7-8 have responses viewable for the student but not instructors
+        assertEquals(6, bundle.getQuestionsNotVisibleForPreviewMap().size());
+        assertEquals(0, bundle.getQuestionsWithCommentNotVisibleForPreview().size());
     }
 
     @Test
@@ -1384,7 +1435,7 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
         InstructorAttributes instructor = responseBundle.instructors.get("instructor1OfCourse1");
         SessionResultsBundle bundle = frLogic.getSessionResultsForUser(
                 session.getFeedbackSessionName(), session.getCourseId(), instructor.getEmail(),
-                true, null);
+                true, null, false);
 
         // Instructor can see responses: q3r1, q6r1
         // We don't check the actual IDs as this is also implicitly tested
@@ -1421,6 +1472,10 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
         // no entry in comment visibility table
         Map<Long, Boolean> commentGiverVisibilityTable = bundle.getCommentGiverVisibilityTable();
         assertEquals(0, commentGiverVisibilityTable.size());
+
+        // preview invisibility info should be empty because this is NOT previewing results
+        assertEquals(0, bundle.getQuestionsNotVisibleForPreviewMap().size());
+        assertEquals(0, bundle.getQuestionsWithCommentNotVisibleForPreview().size());
     }
 
     @Test
@@ -1650,7 +1705,7 @@ public class FeedbackResponsesLogicTest extends BaseLogicTest {
 
         SessionResultsBundle bundle = frLogic.getSessionResultsForUser(
                 fq.getFeedbackSessionName(), fq.getCourseId(), student.getEmail(),
-                false, fq.getId());
+                false, fq.getId(), false);
         assertEquals(1, bundle.getQuestionResponseMap().size());
         List<FeedbackResponseAttributes> responseForQuestion =
                 bundle.getQuestionResponseMap().entrySet().iterator().next().getValue();
