@@ -119,6 +119,9 @@ export class QuestionEditFormComponent {
   @Input()
   isDisplayOnly: boolean = false;
 
+  @Input()
+  isQuestionPublished: boolean = false;
+
   model: QuestionEditFormModel = {
     feedbackQuestionId: '',
 
@@ -281,7 +284,7 @@ export class QuestionEditFormComponent {
       const doChangesNeedWarning: boolean = this.model.isQuestionDetailsChanged
         || this.model.isVisibilityChanged
         || this.model.isFeedbackPathChanged;
-      if (!this.model.isQuestionHasResponses || !doChangesNeedWarning) {
+      if (!this.isQuestionPublished && (!this.model.isQuestionHasResponses || !doChangesNeedWarning)) {
         this.saveExistingQuestionEvent.emit();
       } else if (this.model.isFeedbackPathChanged) {
         // warn user that editing feedback path will delete all messages
