@@ -37,6 +37,7 @@ export class RubricQuestionStatisticsCalculation
   hasWeights: boolean = false;
   weights: number[][] = [];
   answers: number[][] = [];
+  isWeightStatsVisible: boolean = false;
 
   percentages: number[][] = [];
   subQuestionWeightAverage: number[] = [];
@@ -59,6 +60,8 @@ export class RubricQuestionStatisticsCalculation
     this.choices = this.question.rubricChoices;
     this.hasWeights = this.question.hasAssignedWeights;
     this.weights = this.question.rubricWeightsForEachCell;
+    this.isWeightStatsVisible =
+      this.hasWeights && this.weights.length > 0 && this.weights[0].length > 0;
 
     const emptyAnswers: number[][] = [];
     for (let i = 0; i < this.question.rubricSubQuestions.length; i += 1) {
@@ -89,7 +92,7 @@ export class RubricQuestionStatisticsCalculation
     this.percentagesExcludeSelf = this.calculatePercentages(this.answersExcludeSelf);
 
     // only apply weights average if applicable
-    if (!this.hasWeights) {
+    if (!this.isWeightStatsVisible) {
       return;
     }
 
@@ -98,25 +101,28 @@ export class RubricQuestionStatisticsCalculation
 
     // calculate per recipient stats
     for (const response of this.responses) {
-      this.perRecipientStatsMap[response.recipient] = this.perRecipientStatsMap[response.recipient] || {
-        recipientName: response.recipient,
-        recipientEmail: response.recipientEmail,
-        recipientTeam: response.recipientTeam,
-        answers: JSON.parse(JSON.stringify(emptyAnswers)),
-        answersSum: [],
-        percentages: [],
-        percentagesAverage: [],
-        weightsAverage: [],
-        subQuestionTotalChosenWeight: this.subQuestions.map(() => 0),
-        subQuestionWeightAverage: [],
-      };
+      this.perRecipientStatsMap[response.recipientEmail || response.recipient] =
+        this.perRecipientStatsMap[
+          response.recipientEmail || response.recipient
+        ] || {
+          recipientName: response.recipient,
+          recipientEmail: response.recipientEmail,
+          recipientTeam: response.recipientTeam,
+          answers: JSON.parse(JSON.stringify(emptyAnswers)),
+          answersSum: [],
+          percentages: [],
+          percentagesAverage: [],
+          weightsAverage: [],
+          subQuestionTotalChosenWeight: this.subQuestions.map(() => 0),
+          subQuestionWeightAverage: [],
+        };
       for (let i: number = 0; i < response.responseDetails.answer.length; i += 1) {
         const subAnswer: number = response.responseDetails.answer[i];
         if (subAnswer === RUBRIC_ANSWER_NOT_CHOSEN) {
           continue;
         }
-        this.perRecipientStatsMap[response.recipient].answers[i][subAnswer] += 1;
-        this.perRecipientStatsMap[response.recipient].subQuestionTotalChosenWeight[i] +=
+        this.perRecipientStatsMap[response.recipientEmail || response.recipient].answers[i][subAnswer] += 1;
+        this.perRecipientStatsMap[response.recipientEmail || response.recipient].subQuestionTotalChosenWeight[i] +=
             +this.weights[i][subAnswer].toFixed(5);
       }
     }
@@ -124,18 +130,19 @@ export class RubricQuestionStatisticsCalculation
     for (const recipient of Object.keys(this.perRecipientStatsMap)) {
       const perRecipientStats: PerRecipientStats = this.perRecipientStatsMap[recipient];
 
+      // Answers sum = number of answers in each column
       perRecipientStats.answersSum = this.calculateAnswersSum(perRecipientStats.answers);
       perRecipientStats.percentages = this.calculatePercentages(perRecipientStats.answers);
-      perRecipientStats.percentagesAverage =
-          this.calculatePercentagesAverage(perRecipientStats.percentages,
-          this.calculateNumResponses(perRecipientStats.answersSum));
+      perRecipientStats.percentagesAverage = this.calculatePercentagesAverage(perRecipientStats.answersSum);
       perRecipientStats.subQuestionWeightAverage =
           this.calculateSubQuestionWeightAverage(perRecipientStats.answers);
       perRecipientStats.weightsAverage = this.calculateWeightsAverage(this.weights);
-      perRecipientStats.overallWeightedSum = this.calculateWeightedSum(perRecipientStats.percentagesAverage,
-          perRecipientStats.answersSum, perRecipientStats.weightsAverage);
+      // Overall weighted sum = sum of total chosen weight for all sub questions
+      perRecipientStats.overallWeightedSum =
+        +(perRecipientStats.subQuestionTotalChosenWeight.reduce((a, b) => a + b)).toFixed(2);
+      // Overall weighted average = overall weighted sum / total number of responses
       perRecipientStats.overallWeightAverage = +(perRecipientStats.overallWeightedSum
-          / (this.calculateNumResponses(perRecipientStats.answersSum))).toFixed(2);
+          / this.calculateNumResponses(perRecipientStats.answersSum)).toFixed(2);
     }
   }
 
@@ -195,32 +202,19 @@ export class RubricQuestionStatisticsCalculation
   }
 
   // Calculate percentage average for each column
-  private calculatePercentagesAverage(percentages: number[][], numAnswers: number): number[] {
-    // Calculate sum of percentages for each column
-    const sums: number[] = this.calculateAnswersSum(percentages);
+  private calculatePercentagesAverage(answersSum: number[]): number[] {
+    // Calculate total number of responses
+    const numResponses = this.calculateNumResponses(answersSum);
     const averages: number[] = [];
-    // Divide each percentage by the number of answers
-    for (let i: number = 0; i < sums.length; i += 1) {
-      averages[i] = numAnswers === 0 ? 0 : +(sums[i] / numAnswers).toFixed(2);
+    // Divide each column sum by total number of responses, then convert to percentage
+    for (let i: number = 0; i < answersSum.length; i += 1) {
+      averages[i] = numResponses === 0 ? 0 : +(answersSum[i] * 100 / numResponses).toFixed(2);
     }
     return averages;
   }
 
-  // Calculate weighted sum of responses
-  private calculateWeightedSum(percentages: number[], answers: number[], weights: number[]): number {
-    let sum: number = 0;
-    for (let i: number = 0; i < answers.length; i += 1) {
-      sum += (percentages[i] / 100) * answers[i] * weights[i];
-    }
-    return +sum.toFixed(2);
-  }
-
   // Calculate total number of responses
   private calculateNumResponses(answersSum: number[]): number {
-    let num: number = 0;
-    for (let i = 0; i < answersSum.length; i += 1) {
-      num += answersSum[i];
-    }
-    return num;
+    return answersSum.reduce((a, b) => a + b);
   }
 }
