@@ -61,19 +61,25 @@ export abstract class InstructorSessionBasePageComponent {
   protected copyFeedbackSession(fromFeedbackSession: FeedbackSession, newSessionName: string, newCourseId: string,
       oldCourseId: string): Observable<FeedbackSession> {
     // Local constants
-    const twoHoursBeforeNow = moment().tz(fromFeedbackSession.timeZone).subtract(2, 'hours')
+    const startHour = moment.utc(fromFeedbackSession.submissionStartTimestamp).tz(fromFeedbackSession.timeZone).hours();
+    const endHour = moment(fromFeedbackSession.submissionEndTimestamp).tz(fromFeedbackSession.timeZone).hours();
+    const twoHoursBeforeNow = moment().subtract(2, 'hours')
         .valueOf();
-    const twoDaysFromNowRoundedUp = moment().tz(fromFeedbackSession.timeZone).add(2, 'days').startOf('hour')
+    const twoDaysFromNowSameHour = moment().tz(fromFeedbackSession.timeZone).add(2, 'days')
+        .set('hour', startHour)
+        .startOf('hour')
         .valueOf();
-    const sevenDaysFromNowRoundedUp = moment().tz(fromFeedbackSession.timeZone).add(7, 'days').startOf('hour')
+    const sevenDaysFromNowSameHour = moment().tz(fromFeedbackSession.timeZone).add(7, 'days')
+        .set('hour', endHour)
+        .startOf('hour')
         .valueOf();
     const ninetyDaysFromNow = moment().tz(fromFeedbackSession.timeZone).add(90, 'days')
         .valueOf();
-    const ninetyDaysFromNowRoundedUp = moment().tz(fromFeedbackSession.timeZone).add(90, 'days').startOf('hour')
+    const ninetyDaysFromNowRoundedDown = moment().tz(fromFeedbackSession.timeZone).add(90, 'days').startOf('hour')
         .valueOf();
     const oneHundredAndEightyDaysFromNow = moment().tz(fromFeedbackSession.timeZone).add(180, 'days')
         .valueOf();
-    const oneHundredAndEightyDaysFromNowRoundedUp = moment().tz(fromFeedbackSession.timeZone).add(180, 'days')
+    const oneHundredAndEightyDaysFromNowRoundedDown = moment().tz(fromFeedbackSession.timeZone).add(180, 'days')
         .startOf('hour')
         .valueOf();
 
@@ -82,19 +88,19 @@ export abstract class InstructorSessionBasePageComponent {
 
     let copiedSubmissionStartTimestamp = fromFeedbackSession.submissionStartTimestamp;
     if (copiedSubmissionStartTimestamp < twoHoursBeforeNow) {
-      copiedSubmissionStartTimestamp = twoDaysFromNowRoundedUp;
+      copiedSubmissionStartTimestamp = twoDaysFromNowSameHour;
       isModified = true;
     } else if (copiedSubmissionStartTimestamp > ninetyDaysFromNow) {
-      copiedSubmissionStartTimestamp = ninetyDaysFromNowRoundedUp;
+      copiedSubmissionStartTimestamp = ninetyDaysFromNowRoundedDown;
       isModified = true;
     }
 
     let copiedSubmissionEndTimestamp = fromFeedbackSession.submissionEndTimestamp;
     if (copiedSubmissionEndTimestamp < copiedSubmissionStartTimestamp) {
-      copiedSubmissionEndTimestamp = sevenDaysFromNowRoundedUp;
+      copiedSubmissionEndTimestamp = sevenDaysFromNowSameHour;
       isModified = true;
     } else if (copiedSubmissionEndTimestamp > oneHundredAndEightyDaysFromNow) {
-      copiedSubmissionEndTimestamp = oneHundredAndEightyDaysFromNowRoundedUp;
+      copiedSubmissionEndTimestamp = oneHundredAndEightyDaysFromNowRoundedDown;
       isModified = true;
     }
 
@@ -129,7 +135,7 @@ export abstract class InstructorSessionBasePageComponent {
     if (isModified) {
       this.coursesOfModifiedSession.push(newCourseId);
 
-      this.modifiedSession[fromFeedbackSession.feedbackSessionName] = {
+      this.modifiedSession[newSessionName] = {
         oldTimestamp: {
           submissionStartTimestamp: this.formatTimestamp(fromFeedbackSession.submissionStartTimestamp,
               fromFeedbackSession.timeZone),
@@ -151,24 +157,24 @@ export abstract class InstructorSessionBasePageComponent {
       };
 
       if (fromFeedbackSession.responseVisibleSetting === ResponseVisibleSetting.AT_VISIBLE) {
-        this.modifiedSession[fromFeedbackSession.feedbackSessionName].oldTimestamp.responseVisibleTimestamp =
+        this.modifiedSession[newSessionName].oldTimestamp.responseVisibleTimestamp =
             'On session visible time';
       } else if (fromFeedbackSession.responseVisibleSetting === ResponseVisibleSetting.LATER) {
-        this.modifiedSession[fromFeedbackSession.feedbackSessionName].oldTimestamp.responseVisibleTimestamp =
+        this.modifiedSession[newSessionName].oldTimestamp.responseVisibleTimestamp =
             'Not now (publish manually)';
       } else {
-        this.modifiedSession[fromFeedbackSession.feedbackSessionName].oldTimestamp.responseVisibleTimestamp =
+        this.modifiedSession[newSessionName].oldTimestamp.responseVisibleTimestamp =
             this.formatTimestamp(fromFeedbackSession.customResponseVisibleTimestamp!, fromFeedbackSession.timeZone);
       }
 
       if (copiedResponseVisibleSetting === ResponseVisibleSetting.AT_VISIBLE) {
-        this.modifiedSession[fromFeedbackSession.feedbackSessionName].newTimestamp.responseVisibleTimestamp =
+        this.modifiedSession[newSessionName].newTimestamp.responseVisibleTimestamp =
             'On session visible time';
       } else if (copiedResponseVisibleSetting === ResponseVisibleSetting.LATER) {
-        this.modifiedSession[fromFeedbackSession.feedbackSessionName].newTimestamp.responseVisibleTimestamp =
+        this.modifiedSession[newSessionName].newTimestamp.responseVisibleTimestamp =
             'Not now (publish manually)';
       } else {
-        this.modifiedSession[fromFeedbackSession.feedbackSessionName].newTimestamp.responseVisibleTimestamp =
+        this.modifiedSession[newSessionName].newTimestamp.responseVisibleTimestamp =
             this.formatTimestamp(copiedCustomResponseVisibleTimestamp!, fromFeedbackSession.timeZone);
       }
 
@@ -473,7 +479,7 @@ interface SessionTimestampData {
   responseVisibleTimestamp: string;
 }
 
-interface TweakedTimestampData {
+export interface TweakedTimestampData {
   oldTimestamp: SessionTimestampData;
   newTimestamp: SessionTimestampData;
 }
