@@ -5,6 +5,9 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 import { LoadingRetryModule } from '../../components/loading-retry/loading-retry.module';
 import { LoadingSpinnerModule } from '../../components/loading-spinner/loading-spinner.module';
+import {
+  ModifiedTimestampModalModule,
+} from '../../components/modified-timestamps-modal/modified-timestamps-module.module';
 import { SessionEditFormModule } from '../../components/session-edit-form/session-edit-form.module';
 import {
   SessionsRecycleBinTableModule,
@@ -45,6 +48,7 @@ const routes: Routes = [
     LoadingSpinnerModule,
     LoadingRetryModule,
     TeammatesRouterModule,
+    ModifiedTimestampModalModule,
   ],
   declarations: [
     InstructorSessionsPageComponent,
@@ -54,11 +58,6 @@ const routes: Routes = [
   ],
   exports: [
     InstructorSessionsPageComponent,
-  ],
-  entryComponents: [
-    CopyFromOtherSessionsModalComponent,
-    SessionPermanentDeletionConfirmModalComponent,
-    SessionsPermanentDeletionConfirmModalComponent,
   ],
 })
 export class InstructorSessionsPageModule { }
