@@ -477,6 +477,7 @@ export class SessionSubmissionPageComponent implements OnInit, AfterViewInit {
               const model: QuestionSubmissionFormModel = {
                 isLoading: false,
                 isLoaded: false,
+                isTabExpanded: true,
                 feedbackQuestionId: feedbackQuestion.feedbackQuestionId,
 
                 questionNumber: feedbackQuestion.questionNumber,
@@ -731,7 +732,6 @@ export class SessionSubmissionPageComponent implements OnInit, AfterViewInit {
             const isFeedbackResponseDetailsEmpty: boolean =
                 this.feedbackResponsesService.isFeedbackResponseDetailsEmpty(
                     questionSubmissionFormModel.questionType, recipientSubmissionFormModel.responseDetails);
-            isQuestionFullyAnswered = isQuestionFullyAnswered && !isFeedbackResponseDetailsEmpty;
 
             if (!isFeedbackResponseDetailsEmpty) {
               responses.push({
@@ -740,6 +740,8 @@ export class SessionSubmissionPageComponent implements OnInit, AfterViewInit {
               });
             }
           });
+
+      isQuestionFullyAnswered = responses.length > 0;
 
       if (!failToSaveQuestions[questionSubmissionFormModel.questionNumber]) {
         savingRequests.push(
